Fix evince SyncTeX integration scripts
The evince forward and backward search documents
have incorrect file escape configuration which is
causing paths with spaces and other special
characters to not function correctly. Further, the
FindDocument method of the evince dbus daemon does
not support filepaths so I have made the call use
the URI format.

I have also included a missing import for both
versions of the script.